Project 2 for COP3530
Takes in an initial string and hint string from user input, where the
first inputted line is the initial string and the second is a hint
string.
The output of this is a combination of integers. This combination is the
difference in the incides of each occurence of the hint string. In
addition to this, the output must be the longest common subsequence of
numbers.